chore: separate RAP rules by module type
